Create fetch_analytics.scala
1. The project can be run on VS Code, Databricks or jupyter Notebook

2. Spark and Scala are used to analyze two anonymized datasets based on fetch reward receipts and receipts items

3. The idea is to create a modularized code to be able to run reports automatically for 3 analytics ideas based on parameter settings

4. First idea is to generate report for total Sales by BRAND; Here you can see top BRANDS with more walletshare

5. Second is to generate report for sales by State

6. The final idea is to see report for sales by Store Name

All above ideas are design in a way to re-use the same API to create different reports depending parameters you define such as which analysis to run, how many months back to run the analysis, and summary output file to generate for the reports either csv, parquet or json.

Note, codes are commented to easily reason about the logic involved in the process.

The API is design for DATAFRAME API Scenario using Spark/Scala as well as another API for pure Spark SQL.

Thanks